Remove util package, move database.go to auth package
